Fixes for MSVC warnings.

+++ src/maidsafe/routing/tests/routing_functional_test.cc
@@ -482,7 +482,7 @@ TEST_F(RoutingNetworkTest, FUNC_IsConnectedClient) {
   ASSERT_LE(env_->nodes_.size() - env_->ClientIndex(), Parameters::max_non_routing_table_size + 1);
 
   // Vault checks close client id - expect true
-  for (uint16_t i(env_->ClientIndex()); i < env_->nodes_.size(); ++i) {
+  for (size_t i(env_->ClientIndex()); i < env_->nodes_.size(); ++i) {
     NodeId client_id(env_->nodes_.at(i)->node_id());
     std::vector<NodeInfo> closest_nodes(
           env_->GetClosestVaults(client_id, Parameters::max_client_routing_table_size));
